requirements.txt
alembic
barcode
bcrypt
colorlog
cryptography
fastapi
pdfkit
psycopg2-binary
pydantic
pydantic[email]
pydantic-settings
pydantic_core
pymongo
python-dotenv
python-jose
python-multipart
sqlalchemy
uvicorn
wkhtmltopdf